See about a simpler narrative (raw HTML) container, and interactivity.
I'm mostly just learning Vue.js, still, so this seems ugly. Even to me.
I'm also not sure why the "Press me" button isn't working; it correctly
sets the the `accomplished` property to `true` on the right variable
(the proper index of the `steps` Array), but Vue isn't picking this up
and so isn't showing the "Success" button for the user to click.

Too tired now, gonna figure it out later. Maybe ask for some help, too.